add ppos case
